admin and user all api working and code fully set

+++ controllers/contactController.js
@@ -1,36 +1,41 @@
-const ContactService = require("../services/contactService");
+const ContactService = require("../services/ContactService");
 
 class ContactController {
-
   contactService;
 
   constructor() {
-    this.contactService = new ContactService(); 
+    this.contactService = new ContactService();
   }
 
-
-   async getContactLeads(req, res) {
+  async getContactLeads(req, res) {
     try {
-      const leads = await this.contactService.getContactLeads();
+      const uid = req.user.uid;  
+      if (!uid) {
+        return res.json({ success: false, msg: "UID is missing" });
+      }
+      const leads = await this.contactService.getContactLeads(uid);
       res.json({ data: leads, success: true });
     } catch (err) {
       console.error(err);
-      res.json({ success: false, msg: "Server error" });
+      res.json({ success: false, msg: "Server error", err });
     }
   }
 
-   async deleteContactEntry(req, res) {
+  async deleteContactEntry(req, res) {
     try {
       const { id } = req.body;
+      if (!id) {
+        return res.json({ success: false, msg: "ID is required" });
+      }
       await this.contactService.deleteContactEntry(id);
       res.json({ success: true, msg: "Entry was deleted" });
     } catch (err) {
       console.error(err);
-      res.json({ success: false, msg: "Server error" });
+      res.json({ success: false, msg: "Server error", err });
     }
   }
 
-     async submitContactForm(req, res) {
+  async submitContactForm(req, res) {
     try {
       await this.contactService.submitContactForm(req.body);
       res.json({ success: true, msg: 'Your form has been submitted. We will contact you asap' });
@@ -39,7 +44,6 @@ class ContactController {
       console.log(error);
     }
   }
-
 }
 
 module.exports = ContactController;
